Pink Pokémon's shiny palettes edited
Now, the pink that was taken from the pink palette's light color has had its R G and B values all reduced by 10. It's not likely this was the intended palette, but it's less likely that the original shiny palette was ever intended to look the way it did in the demo.